+++ src/vm/runners/builtin_runner/ec_op.rs
@@ -1,48 +1,15 @@
+use std::any::Any;
+
+use num_bigint::BigInt;
+use num_integer::Integer;
+
 use crate::math_utils::{ec_add, ec_double};
 use crate::types::relocatable::{MaybeRelocatable, Relocatable};
-use crate::vm::errors::memory_errors::MemoryError;
 use crate::vm::errors::runner_errors::RunnerError;
-use crate::vm::vm_memory::memory::{Memory, ValidationRule};
+use crate::vm::runners::builtin_runner::BuiltinRunner;
+use crate::vm::vm_memory::memory::Memory;
 use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
 use crate::{bigint, bigint_str};
-use num_bigint::{BigInt, Sign};
-use num_integer::Integer;
-use num_traits::{One, Zero};
-use starknet_crypto::{pedersen_hash, FieldElement};
-use std::any::Any;
-use std::ops::Shl;
-
-pub struct RangeCheckBuiltinRunner {
-    _ratio: BigInt,
-    base: usize,
-    _stop_ptr: Option<Relocatable>,
-    _cells_per_instance: i32,
-    _n_input_cells: i32,
-    _inner_rc_bound: BigInt,
-    pub _bound: BigInt,
-    _n_parts: u32,
-}
-pub struct OutputBuiltinRunner {
-    base: usize,
-    _stop_ptr: Option<Relocatable>,
-}
-
-pub struct HashBuiltinRunner {
-    pub base: usize,
-    _ratio: usize,
-    cells_per_instance: usize,
-    _n_input_cells: usize,
-    _stop_ptr: Option<Relocatable>,
-    verified_addresses: Vec<Relocatable>,
-}
-
-pub struct BitwiseBuiltinRunner {
-    _ratio: usize,
-    pub base: usize,
-    cells_per_instance: usize,
-    _n_input_cells: usize,
-    total_n_bits: u32,
-}
 
 pub struct EcOpBuiltinRunner {
     _ratio: usize,
@@ -52,272 +19,6 @@ pub struct EcOpBuiltinRunner {
     scalar_height: usize,
     _scalar_bits: usize,
     scalar_limit: BigInt,
-}
-
-pub trait BuiltinRunner {
-    ///Creates the necessary segments for the builtin in the MemorySegmentManager and stores the first address on the builtin's base
-    fn initialize_segments(&mut self, segments: &mut MemorySegmentManager, memory: &mut Memory);
-    fn initial_stack(&self) -> Vec<MaybeRelocatable>;
-    ///Returns the builtin's base
-    fn base(&self) -> Relocatable;
-    fn add_validation_rule(&self, memory: &mut Memory);
-    fn deduce_memory_cell(
-        &mut self,
-        address: &Relocatable,
-        memory: &Memory,
-    ) -> Result<Option<MaybeRelocatable>, RunnerError>;
-    fn as_any(&self) -> &dyn Any;
-}
-
-impl RangeCheckBuiltinRunner {
-    pub fn new(ratio: BigInt, n_parts: u32) -> RangeCheckBuiltinRunner {
-        let inner_rc_bound = bigint!(1i32 << 16);
-        RangeCheckBuiltinRunner {
-            _ratio: ratio,
-            base: 0,
-            _stop_ptr: None,
-            _cells_per_instance: 1,
-            _n_input_cells: 1,
-            _inner_rc_bound: inner_rc_bound.clone(),
-            _bound: inner_rc_bound.pow(n_parts),
-            _n_parts: n_parts,
-        }
-    }
-}
-impl BuiltinRunner for RangeCheckBuiltinRunner {
-    fn initialize_segments(&mut self, segments: &mut MemorySegmentManager, memory: &mut Memory) {
-        self.base = segments.add(memory).segment_index
-    }
-
-    fn initial_stack(&self) -> Vec<MaybeRelocatable> {
-        vec![MaybeRelocatable::from((self.base, 0))]
-    }
-
-    fn base(&self) -> Relocatable {
-        Relocatable::from((self.base, 0))
-    }
-
-    fn add_validation_rule(&self, memory: &mut Memory) {
-        let rule: ValidationRule = ValidationRule(Box::new(
-            |memory: &Memory,
-             address: &MaybeRelocatable|
-             -> Result<MaybeRelocatable, MemoryError> {
-                if let Some(MaybeRelocatable::Int(ref num)) = memory.get(address)? {
-                    if &BigInt::zero() <= num && num < &BigInt::one().shl(128u8) {
-                        Ok(address.to_owned())
-                    } else {
-                        Err(MemoryError::NumOutOfBounds)
-                    }
-                } else {
-                    Err(MemoryError::FoundNonInt)
-                }
-            },
-        ));
-        memory.add_validation_rule(self.base, rule);
-    }
-
-    fn deduce_memory_cell(
-        &mut self,
-        _address: &Relocatable,
-        _memory: &Memory,
-    ) -> Result<Option<MaybeRelocatable>, RunnerError> {
-        Ok(None)
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-impl OutputBuiltinRunner {
-    pub fn new() -> OutputBuiltinRunner {
-        OutputBuiltinRunner {
-            base: 0,
-            _stop_ptr: None,
-        }
-    }
-}
-
-impl BuiltinRunner for OutputBuiltinRunner {
-    fn initialize_segments(&mut self, segments: &mut MemorySegmentManager, memory: &mut Memory) {
-        self.base = segments.add(memory).segment_index
-    }
-
-    fn initial_stack(&self) -> Vec<MaybeRelocatable> {
-        vec![MaybeRelocatable::from((self.base, 0))]
-    }
-
-    fn base(&self) -> Relocatable {
-        Relocatable::from((self.base, 0))
-    }
-
-    fn add_validation_rule(&self, _memory: &mut Memory) {}
-
-    fn deduce_memory_cell(
-        &mut self,
-        _address: &Relocatable,
-        _memory: &Memory,
-    ) -> Result<Option<MaybeRelocatable>, RunnerError> {
-        Ok(None)
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-impl HashBuiltinRunner {
-    pub fn new(ratio: usize) -> Self {
-        HashBuiltinRunner {
-            base: 0,
-
-            _ratio: ratio,
-            cells_per_instance: 3,
-            _n_input_cells: 2,
-            _stop_ptr: None,
-            verified_addresses: Vec::new(),
-        }
-    }
-}
-
-impl BuiltinRunner for HashBuiltinRunner {
-    fn initialize_segments(&mut self, segments: &mut MemorySegmentManager, memory: &mut Memory) {
-        self.base = segments.add(memory).segment_index
-    }
-
-    fn initial_stack(&self) -> Vec<MaybeRelocatable> {
-        vec![MaybeRelocatable::from((self.base, 0))]
-    }
-
-    fn base(&self) -> Relocatable {
-        Relocatable::from((self.base, 0))
-    }
-
-    fn add_validation_rule(&self, _memory: &mut Memory) {}
-
-    fn deduce_memory_cell(
-        &mut self,
-        address: &Relocatable,
-        memory: &Memory,
-    ) -> Result<Option<MaybeRelocatable>, RunnerError> {
-        if address.offset.mod_floor(&self.cells_per_instance) != 2
-            || self.verified_addresses.contains(address)
-        {
-            return Ok(None);
-        };
-        if let (Ok(Some(MaybeRelocatable::Int(num_a))), Ok(Some(MaybeRelocatable::Int(num_b)))) = (
-            memory.get(&MaybeRelocatable::RelocatableValue(Relocatable {
-                segment_index: address.segment_index,
-                offset: address.offset - 1,
-            })),
-            memory.get(&MaybeRelocatable::RelocatableValue(Relocatable {
-                segment_index: address.segment_index,
-                offset: address.offset - 2,
-            })),
-        ) {
-            self.verified_addresses.push(address.clone());
-
-            //Convert MaybeRelocatable to FieldElement
-            let a_string = num_a.to_str_radix(10);
-            let b_string = num_b.to_str_radix(10);
-            let (y, x) = match (
-                FieldElement::from_dec_str(&a_string),
-                FieldElement::from_dec_str(&b_string),
-            ) {
-                (Ok(field_element_a), Ok(field_element_b)) => (field_element_a, field_element_b),
-                _ => return Err(RunnerError::FailedStringConversion),
-            };
-            //Compute pedersen Hash
-            let fe_result = pedersen_hash(&x, &y);
-            //Convert result from FieldElement to MaybeRelocatable
-            let r_byte_slice = fe_result.to_bytes_be();
-            let result = BigInt::from_bytes_be(Sign::Plus, &r_byte_slice);
-            return Ok(Some(MaybeRelocatable::from(result)));
-        }
-        Ok(None)
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-impl BitwiseBuiltinRunner {
-    pub fn new(ratio: usize) -> Self {
-        BitwiseBuiltinRunner {
-            base: 0,
-
-            _ratio: ratio,
-            cells_per_instance: 5,
-            _n_input_cells: 2,
-            total_n_bits: 251,
-        }
-    }
-}
-
-impl BuiltinRunner for BitwiseBuiltinRunner {
-    fn initialize_segments(&mut self, segments: &mut MemorySegmentManager, memory: &mut Memory) {
-        self.base = segments.add(memory).segment_index
-    }
-
-    fn initial_stack(&self) -> Vec<MaybeRelocatable> {
-        vec![MaybeRelocatable::from((self.base, 0))]
-    }
-
-    fn base(&self) -> Relocatable {
-        Relocatable::from((self.base, 0))
-    }
-
-    fn add_validation_rule(&self, _memory: &mut Memory) {}
-
-    fn deduce_memory_cell(
-        &mut self,
-        address: &Relocatable,
-        memory: &Memory,
-    ) -> Result<Option<MaybeRelocatable>, RunnerError> {
-        let index = address.offset.mod_floor(&self.cells_per_instance);
-        if index == 0 || index == 1 {
-            return Ok(None);
-        }
-        let x_addr = MaybeRelocatable::from((address.segment_index, address.offset - index));
-        let y_addr = x_addr.add_usize_mod(1, None);
-        if let (Ok(Some(MaybeRelocatable::Int(num_x))), Ok(Some(MaybeRelocatable::Int(num_y)))) =
-            (memory.get(&x_addr), memory.get(&y_addr))
-        {
-            let _2_pow_bits = bigint!(1).shl(self.total_n_bits);
-            if num_x >= &_2_pow_bits {
-                return Err(RunnerError::IntegerBiggerThanPowerOfTwo(
-                    x_addr,
-                    self.total_n_bits,
-                    num_x.clone(),
-                ));
-            };
-            if num_y >= &_2_pow_bits {
-                return Err(RunnerError::IntegerBiggerThanPowerOfTwo(
-                    y_addr,
-                    self.total_n_bits,
-                    num_y.clone(),
-                ));
-            };
-            let res = match index {
-                2 => Some(MaybeRelocatable::from(num_x & num_y)),
-                3 => Some(MaybeRelocatable::from(num_x ^ num_y)),
-                4 => Some(MaybeRelocatable::from(num_x | num_y)),
-                _ => None,
-            };
-            return Ok(res);
-        }
-        Ok(None)
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-impl Default for OutputBuiltinRunner {
-    fn default() -> Self {
-        Self::new()
-    }
 }
 
 impl EcOpBuiltinRunner {
@@ -479,129 +180,7 @@ impl BuiltinRunner for EcOpBuiltinRunner {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{bigint, bigint_str, utils::test_utils::*, vm::vm_memory::memory::Memory};
-
-    #[test]
-    fn initialize_segments_for_output() {
-        let mut builtin = OutputBuiltinRunner::new();
-        let mut segments = MemorySegmentManager::new();
-        let mut memory = Memory::new();
-        builtin.initialize_segments(&mut segments, &mut memory);
-        assert_eq!(builtin.base, 0);
-    }
-
-    #[test]
-    fn initialize_segments_for_range_check() {
-        let mut builtin = RangeCheckBuiltinRunner::new(bigint!(8), 8);
-        let mut segments = MemorySegmentManager::new();
-        let mut memory = Memory::new();
-        builtin.initialize_segments(&mut segments, &mut memory);
-        assert_eq!(builtin.base, 0);
-    }
-
-    #[test]
-    fn get_initial_stack_for_range_check_with_base() {
-        let mut builtin = RangeCheckBuiltinRunner::new(bigint!(8), 8);
-        builtin.base = 1;
-        let initial_stack = builtin.initial_stack();
-        assert_eq!(
-            initial_stack[0].clone(),
-            MaybeRelocatable::RelocatableValue(builtin.base())
-        );
-        assert_eq!(initial_stack.len(), 1);
-    }
-
-    #[test]
-    fn get_initial_stack_for_output_with_base() {
-        let mut builtin = OutputBuiltinRunner::new();
-        builtin.base = 1;
-        let initial_stack = builtin.initial_stack();
-        assert_eq!(
-            initial_stack[0].clone(),
-            MaybeRelocatable::RelocatableValue(builtin.base())
-        );
-        assert_eq!(initial_stack.len(), 1);
-    }
-
-    #[test]
-    fn deduce_memory_cell_pedersen_for_preset_memory_valid() {
-        let memory = memory![((0, 3), 32), ((0, 4), 72), ((0, 5), 0)];
-        let mut builtin = HashBuiltinRunner::new(8);
-
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 5)), &memory);
-        assert_eq!(
-            result,
-            Ok(Some(MaybeRelocatable::from(bigint_str!(
-                b"3270867057177188607814717243084834301278723532952411121381966378910183338911"
-            ))))
-        );
-        assert_eq!(builtin.verified_addresses, vec![Relocatable::from((0, 5))]);
-    }
-
-    #[test]
-    fn deduce_memory_cell_pedersen_for_preset_memory_incorrect_offset() {
-        let memory = memory![((0, 4), 32), ((0, 5), 72), ((0, 6), 0)];
-        let mut builtin = HashBuiltinRunner::new(8);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 6)), &memory);
-        assert_eq!(result, Ok(None));
-    }
-
-    #[test]
-    fn deduce_memory_cell_pedersen_for_preset_memory_no_values_to_hash() {
-        let memory = memory![((0, 4), 72), ((0, 5), 0)];
-        let mut builtin = HashBuiltinRunner::new(8);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 5)), &memory);
-        assert_eq!(result, Ok(None));
-    }
-
-    #[test]
-    fn deduce_memory_cell_pedersen_for_preset_memory_already_computed() {
-        let memory = memory![((0, 3), 32), ((0, 4), 72), ((0, 5), 0)];
-        let mut builtin = HashBuiltinRunner::new(8);
-        builtin.verified_addresses = vec![Relocatable::from((0, 5))];
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 5)), &memory);
-        assert_eq!(result, Ok(None));
-    }
-
-    #[test]
-    fn deduce_memory_cell_bitwise_for_preset_memory_valid_and() {
-        let memory = memory![((0, 5), 10), ((0, 6), 12), ((0, 7), 0)];
-        let mut builtin = BitwiseBuiltinRunner::new(256);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 7)), &memory);
-        assert_eq!(result, Ok(Some(MaybeRelocatable::from(bigint!(8)))));
-    }
-
-    #[test]
-    fn deduce_memory_cell_bitwise_for_preset_memory_valid_xor() {
-        let memory = memory![((0, 5), 10), ((0, 6), 12), ((0, 8), 0)];
-        let mut builtin = BitwiseBuiltinRunner::new(256);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 8)), &memory);
-        assert_eq!(result, Ok(Some(MaybeRelocatable::from(bigint!(6)))));
-    }
-
-    #[test]
-    fn deduce_memory_cell_bitwise_for_preset_memory_valid_or() {
-        let memory = memory![((0, 5), 10), ((0, 6), 12), ((0, 9), 0)];
-        let mut builtin = BitwiseBuiltinRunner::new(256);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 9)), &memory);
-        assert_eq!(result, Ok(Some(MaybeRelocatable::from(bigint!(14)))));
-    }
-
-    #[test]
-    fn deduce_memory_cell_bitwise_for_preset_memory_incorrect_offset() {
-        let memory = memory![((0, 3), 10), ((0, 4), 12), ((0, 5), 0)];
-        let mut builtin = BitwiseBuiltinRunner::new(256);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 5)), &memory);
-        assert_eq!(result, Ok(None));
-    }
-
-    #[test]
-    fn deduce_memory_cell_bitwise_for_preset_memory_no_values_to_operate() {
-        let memory = memory![((0, 5), 12), ((0, 7), 0)];
-        let mut builtin = BitwiseBuiltinRunner::new(256);
-        let result = builtin.deduce_memory_cell(&Relocatable::from((0, 5)), &memory);
-        assert_eq!(result, Ok(None));
-    }
+    use crate::utils::test_utils::memory;
 
     #[test]
     fn point_is_on_curve_a() {
