+++ src/vm/runners/builtin_runner/range_check.rs
@@ -1,0 +1,110 @@
+use std::any::Any;
+use std::ops::Shl;
+
+use num_bigint::BigInt;
+use num_traits::{One, Zero};
+
+use crate::bigint;
+use crate::types::relocatable::{MaybeRelocatable, Relocatable};
+use crate::vm::errors::memory_errors::MemoryError;
+use crate::vm::errors::runner_errors::RunnerError;
+use crate::vm::runners::builtin_runner::BuiltinRunner;
+use crate::vm::vm_memory::memory::{Memory, ValidationRule};
+use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
+
+pub struct RangeCheckBuiltinRunner {
+    _ratio: BigInt,
+    base: usize,
+    _stop_ptr: Option<Relocatable>,
+    _cells_per_instance: i32,
+    _n_input_cells: i32,
+    _inner_rc_bound: BigInt,
+    pub _bound: BigInt,
+    _n_parts: u32,
+}
+
+impl RangeCheckBuiltinRunner {
+    pub fn new(ratio: BigInt, n_parts: u32) -> RangeCheckBuiltinRunner {
+        let inner_rc_bound = bigint!(1i32 << 16);
+        RangeCheckBuiltinRunner {
+            _ratio: ratio,
+            base: 0,
+            _stop_ptr: None,
+            _cells_per_instance: 1,
+            _n_input_cells: 1,
+            _inner_rc_bound: inner_rc_bound.clone(),
+            _bound: inner_rc_bound.pow(n_parts),
+            _n_parts: n_parts,
+        }
+    }
+}
+impl BuiltinRunner for RangeCheckBuiltinRunner {
+    fn initialize_segments(&mut self, segments: &mut MemorySegmentManager, memory: &mut Memory) {
+        self.base = segments.add(memory).segment_index
+    }
+
+    fn initial_stack(&self) -> Vec<MaybeRelocatable> {
+        vec![MaybeRelocatable::from((self.base, 0))]
+    }
+
+    fn base(&self) -> Relocatable {
+        Relocatable::from((self.base, 0))
+    }
+
+    fn add_validation_rule(&self, memory: &mut Memory) {
+        let rule: ValidationRule = ValidationRule(Box::new(
+            |memory: &Memory,
+             address: &MaybeRelocatable|
+             -> Result<MaybeRelocatable, MemoryError> {
+                if let Some(MaybeRelocatable::Int(ref num)) = memory.get(address)? {
+                    if &BigInt::zero() <= num && num < &BigInt::one().shl(128u8) {
+                        Ok(address.to_owned())
+                    } else {
+                        Err(MemoryError::NumOutOfBounds)
+                    }
+                } else {
+                    Err(MemoryError::FoundNonInt)
+                }
+            },
+        ));
+        memory.add_validation_rule(self.base, rule);
+    }
+
+    fn deduce_memory_cell(
+        &mut self,
+        _address: &Relocatable,
+        _memory: &Memory,
+    ) -> Result<Option<MaybeRelocatable>, RunnerError> {
+        Ok(None)
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn initialize_segments_for_range_check() {
+        let mut builtin = RangeCheckBuiltinRunner::new(bigint!(8), 8);
+        let mut segments = MemorySegmentManager::new();
+        let mut memory = Memory::new();
+        builtin.initialize_segments(&mut segments, &mut memory);
+        assert_eq!(builtin.base, 0);
+    }
+
+    #[test]
+    fn get_initial_stack_for_range_check_with_base() {
+        let mut builtin = RangeCheckBuiltinRunner::new(bigint!(8), 8);
+        builtin.base = 1;
+        let initial_stack = builtin.initial_stack();
+        assert_eq!(
+            initial_stack[0].clone(),
+            MaybeRelocatable::RelocatableValue(builtin.base())
+        );
+        assert_eq!(initial_stack.len(), 1);
+    }
+}
